refactor: restructure articles page and update ArticleList component
