+++ app/_features/articles/ArticleList.tsx
@@ -1,3 +1,4 @@
+import { UniversalLink } from "../viewTransition";
 import { ArticleListItem } from "./ArticleListItem";
 import { getArticlesMeta } from "./getArticlesMeta";
 import type { TagEnum } from "./type";
@@ -12,10 +13,16 @@ export const ArticleList = ({ tag }: Props): JSX.Element => {
   const articlesMeta = getArticlesMeta({ tag });
 
   return (
-    <div className="grid gap-12">
-      {articlesMeta.map((article) => (
-        <ArticleListItem key={article.title} {...article} />
-      ))}
+    <div className="blur-enter-content">
+      <UniversalLink href="/" isEnabledUnderline className="text-zinc-700 dark:text-zinc-300">
+        ← Go back
+      </UniversalLink>
+
+      <div className="mt-12 grid gap-6 blur-enter-content enter-step-150">
+        {articlesMeta.map((article) => (
+          <ArticleListItem key={article.title} {...article} />
+        ))}
+      </div>
     </div>
   );
 };
