+++ app/globals.css
@@ -112,6 +112,10 @@
       animation-delay: calc(var(--enter-initial) + var(--enter-stage) * var(--enter-step));
     }
 
+    .blur-enter-content.enter-step-150 > * {
+      --enter-step: 150ms;
+    }
+
     .blur-enter-content > *:nth-child(1) {
       --enter-stage: 1;
     }
